requirements/base.txt
six
PyYAML